Refactored to classical component, rendered sub functional component
